implementacion swagger

+++ src/main/java/com/aulamatriz/product/dto/ProductDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
@@ -11,7 +13,8 @@ public class ProductDTO {
 
     private String name;
     private String description;
-    private double price;
+    private
+    BigDecimal price;
     private int stock;
